sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightningrod"
version = "0.1.0"
description = "Python SDK for Lightning Rod AI-powered forecasting dataset generation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    {name = "Lightning Rod Labs", email = "support@lightningrod.ai"}
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "attrs>=23.1.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "openapi-python-client>=0.15.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://lightningrod.ai"
Documentation = "https://lightningrod.ai/sdk"
Repository = "https://github.com/lightning-rod-labs/lightningrod-python"


